Change terms accumulator to UINT64 to prevent overflow when building large collections (such as gov2 mainbody field) in a field's total term count.
git-svn-id: svn+ssh://svn.code.sf.net/p/lemur/code/indri/trunk@690 14020d50-3050-45fe-8e0c-5ed7005ca764

+++ src/IndexWriter.cpp
@@ -320,7 +320,7 @@ void IndexWriter::_writeFieldList( const std::string& fileName, int fieldIndex, 
   int lastDocument = 0;
 
   int documents = 0;
-  int terms = 0;
+  UINT64 terms = 0;
 
   for( int i=0; i<iterators.size(); i++ ) {
     DocExtentListIterator* iterator = iterators[i];
